requirements.txt
asgiref==3.2.10
Babel==2.8.0
certifi==2020.6.20
chardet==3.0.4
defusedxml==0.6.0
Django==3.0.8
django-allauth==0.42.0
django-countries==6.1.2
django-crispy-forms==1.9.2
django-localflavor==3.0.1
django-phone-field==1.8.1
django-phonenumber-field==4.0.0
idna==2.10
oauthlib==3.1.0
phonenumbers==8.12.7
python-stdnum==1.13
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
sqlparse==0.3.1
stripe==2.49.0
urllib3==1.25.10
